fix compile

+++ iotdb-core/datanode/src/main/java/org/apache/iotdb/db/storageengine/dataregion/tsfile/generator/TsFileNameGenerator.java
@@ -341,7 +341,7 @@ public class TsFileNameGenerator {
                       tsFileResources.get(0).getTimePartition(),
                       tsFileName.time,
                       tsFileName.version,
-                      maxInnerCompactionCount,
+                      (int) maxInnerCompactionCount,
                       (int) maxCrossMergeCount,
                       maxTierLevel,
                       IoTDBConstant.INNER_COMPACTION_TMP_FILE_SUFFIX)),
